App secured from unauth access

+++ app/api/issues/[id]/route.ts
@@ -1,9 +1,15 @@
 
 import schema from "@/app/validationSchemas";
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function PATCH( request: NextRequest, { params }: { params: { id: string } }) {
+
+    const session = await getServerSession();
+    if(!session)
+        return NextResponse.json({}, { status: 401 })
+
     const body = await request.json();
     const validation = schema.safeParse(body);
     
@@ -38,6 +44,9 @@ export async function PATCH( request: NextRequest, { params }: { params: { id: s
 
 
 export async function DELETE( request: NextRequest, { params }: { params: { id: string } }) {
+    const session = await getServerSession();
+    if(!session)
+        return NextResponse.json({}, { status: 401 })
     
     const issue = await prisma.issue.delete({
         where: { id: parseInt(params.id) }
